pyt/requirements.txt
openai
requests
PyYAML
pydub
elevenlabs
pymongo
fastapi>=0.103.1
uvicorn>=0.23.2
python-multipart>=0.0.6
numpy>=1.24.3
sounddevice>=0.4.6
websockets>=11.0.3